Airodump-ng: fixed warnings withs isspace() on cygwin.
git-svn-id: http://svn.aircrack-ng.org/trunk@2396 28c6078b-6c39-48e3-add9-af49d547ecab

+++ src/airodump-ng.c
@@ -388,8 +388,8 @@ void trim(char *str)
     int begin = 0;
     int end = strlen(str) - 1;
 
-    while (isspace(str[begin])) begin++;
-    while ((end >= begin) && isspace(str[end])) end--;
+    while (isspace((int)str[begin])) begin++;
+    while ((end >= begin) && isspace((int)str[end])) end--;
     // Shift all characters back to the start of the string array.
     for (i = begin; i <= end; i++)
         str[i - begin] = str[i];
